requirements.txt
pillow
pyperclip
imagekitio
cryptography
pystray
python-dotenv
google-generativeai
pywin32
olefile